break module out into files and fix compilation errors.
